moved utils into sub folder and extracted general utils
