--- alliance-scanner.ps1
@@ -1,4 +1,0 @@
-python -m venv ./venv
-invoke-expression -Command .\venv\Scripts\Activate.ps1
-pip install -r requirements.txt
-python .\alliance-scanner.py
