--- install_scanner.bat
@@ -1,3 +1,0 @@
-@ECHO OFF
-powershell.exe -noLogo -ExecutionPolicy unrestricted -file %scriptdir%install_scanner.ps1
-PAUSE
