+++ scanner_utils/alliance_mode.py
@@ -1,0 +1,68 @@
+from dummy_root import get_app_root
+
+mode_data = {
+    "alliance": {
+        "name": (334, 260, 293, 33),
+        "score": (1117, 265, 250, 33),
+        "threshold": 90,
+        "invert": True,
+        "script": str(get_app_root() / "deps/inputs/alliance_6_person_scroll.txt"),
+        "normal": {
+            "name_pos": [
+                (334, 260, 293, 33),
+                (334, 359, 293, 33),
+                (334, 460, 293, 33),
+                (334, 562, 293, 33),
+                (334, 662, 293, 33),
+                (334, 763, 293, 33),
+            ],
+            "score_pos": [
+                (1117, 265, 250, 33),
+                (1117, 364, 250, 33),
+                (1117, 465, 250, 33),
+                (1117, 567, 250, 33),
+                (1117, 667, 250, 33),
+                (1117, 768, 250, 33),
+            ],
+        },
+        "last": {
+            "name_pos": [
+                (334, 283, 293, 33),
+                (334, 383, 293, 33),
+                (334, 483, 293, 33),
+                (334, 595, 293, 33),
+                (334, 685, 293, 33),
+                (334, 785, 293, 33),
+            ],
+            "score_pos": [
+                (1117, 288, 250, 33),
+                (1117, 388, 250, 33),
+                (1117, 488, 250, 33),
+                (1117, 590, 250, 33),
+                (1117, 690, 250, 33),
+                (1117, 790, 250, 33),
+            ],
+        },
+    },
+    "honor": {
+        "name": (774, 330, 257, 33),
+        "score": (1183, 330, 178, 33),
+        "threshold": 150,
+        "invert": False,
+        "script": str(get_app_root() / "deps/inputs/honor_5_person_scroll.txt"),
+        "name_pos": [
+            (774, 324, 257, 40),
+            (774, 424, 257, 40),
+            (774, 524, 257, 40),
+            (774, 624, 257, 40),
+            (774, 724, 257, 40),
+        ],
+        "score_pos": [
+            (1183, 324, 178, 40),
+            (1183, 424, 178, 40),
+            (1183, 524, 178, 40),
+            (1183, 624, 178, 40),
+            (1183, 724, 178, 40),
+        ],
+    },
+}
