+++ scanner_utils/validator.py
@@ -2,21 +2,11 @@ import os
 import sys
 import glob
 from pathlib import Path
-
-
-def get_root_dir() -> Path:
-    if getattr(sys, "frozen", False):
-        # If the application is run as a bundle, the PyInstaller bootloader
-        # extends the sys module by a flag frozen=True and sets the app
-        # path into variable _MEIPASS'.
-        print("Bundle detected!")
-        return Path(sys.executable).parent
-    else:
-        return Path(__file__).parent
+from dummy_root import get_app_root
 
 
 def validate_installation() -> bool:
-    root_dir = get_root_dir()
+    root_dir = get_app_root()
 
     tess_dir = root_dir / "deps" / "tessdata"
     adb_dir = root_dir / "deps" / "platform-tools"
